backend/requirements.txt
fastapi
uvicorn
pydantic
sentence-transformers
faiss-cpu
numpy
streamlit
